+++ R/ef_hdv_speed.R
@@ -12,13 +12,80 @@
 #' @param eu Euro emission standard: "PRE", "I", "II", "III", "IV" and "V"
 #' @param gr Gradient or slope of road: -0.06, -0.04, -0.02, 0.00, 0.02. 0.04 or 0.06
 #' @param l Load of the vehicle: 0.0, 0.5 or 1.0
-#' @param p Pollutant: "CO", "FC", "NOx" or "HC"
+#' @param p Character; pollutant: "CO", "FC", "NOx", "HC", "PM", "NMHC", "CH4",
+#' "CO2",  "SO2" or "Pb". Only when p is "SO2" pr "Pb" x is needed. Also
+#' polycyclic aromatic hydrocarbons (PAHs) and persistent organi pollutants (POPs).
+#' @param x Numeric; if pollutant is "SO2", it is sulphur in fuel in ppm, if is
+#' "Pb", Lead in fuel in ppm.
 #' @param k Multiplication factor
 #' @param show.equation Option to see or not the equation parameters
 #' @return an emission factor function which depends of the average speed V  g/km
 #' @keywords speed emission factors
+#' @note  \strong{Pollutants}: "CO", "NOx", "HC", "PM", "CH4", "NMHC", "CO2", "SO2",
+#' "Pb".
+#'
+#' \strong{PAH and POP}: "indeno(1,2,3-cd)pyrene", "benzo(k)fluoranthene",
+#' "benzo(b)fluoranthene", "benzo(ghi)perylene", "fluoranthene",
+#' "benzo(a)pyrene", "pyrene", "perylene",  "anthanthrene", "benzo(b)fluorene",
+#' "benzo(e)pyrene", "triphenylene", "benzo(j)fluoranthene",
+#' "dibenzo(a,j)anthacene", "dibenzo(a,l)pyrene", "3,6-dimethyl-phenanthrene",
+#' "benzo(a)anthracene", "acenaphthylene", "acenapthene", "fluorene",
+#' "chrysene", "phenanthrene", "napthalene",  "anthracene", "coronene",
+#' "dibenzo(ah)anthracene".
+#'
+#' \strong{Dioxins and furans}: PCDD, PCDF and PCB expressed as (g equivalent
+#' toxicity / km).
+#'
+#' \strong{Metals}: "As", "Cd", "Cr", "Cu", "Hg", "Ni", "Pb", "Se", "Zn" (g/km).
+#' \strong{NMHC}:
+#'
+#' \emph{ALKANES}: "ethane", "propane", "butane", "isobutane", "pentane",
+#' "isopentane", "hexane", "heptane", "octane", "TWO_methylhexane", "nonane",
+#' "TWO_methylheptane", "THREE_methylhexane", "decane", "THREE_methylheptane",
+#' "alcanes_C10_C12", "alkanes_C13".
+#'
+#' \emph{CYCLOALKANES}: "cycloalcanes".
+#'
+#' \emph{ALKENES}: "ethylene", "propylene", "propadiene", "ONE_butene",
+#' "isobutene", "TWO_butene", "ONE_3_butadiene", "ONE_pentene", "TWO_pentene",
+#' "ONE_hexene", "dimethylhexene".
+#'
+#' \emph{ALKYNES}:"ONE_butine", "propine", "acetylene".
+#'
+#' \emph{ALDEHYDES}: "formaldehyde", "acetaldehyde", "acrolein", "benzaldehyde",
+#' "crotonaldehyde", "methacrolein", "butyraldehyde", "isobutanaldehyde",
+#' "propionaldehyde", "hexanal", "i_valeraldehyde", "valeraldehyde",
+#' "o_tolualdehyde", "m_tolualdehyde", "p_tolualdehyde".
+#'
+#' \emph{KETONES}: "acetone", "methylethlketone".
+#'
+#' \emph{AROMATICS}: "toluene", "ethylbenzene", "m_p_xylene", "o_xylene",
+#' "ONE_2_3_trimethylbenzene", "ONE_2_4_trimethylbenzene",
+#' "ONE_3_5_trimethylbenzene", "styrene", "benzene", "C9", "C10", "C13".
 #' @export
 #' @examples {
+#' # Quick view
+#' pol <- c("CO", "NOx", "HC", "NMHC", "CH4", "FC", "PM", "CO2", "SO2")
+#' f <- sapply(1:length(pol), function(i){
+#' print(pol[i])
+#' ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
+#' l = 0.5, p = pol[i], x = 10)(30)
+#' })
+#' f
+#' # PAH POP
+#' ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
+#' l = 0.5, p = "napthalene", x = 10)(30)
+#' ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
+#' l = 0.5, p = "fluoranthene", x = 10)(30)
+#'
+#' # Dioxins and Furans
+#' ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
+#' l = 0.5, p = "PCB", x = 10)(30)
+#'
+#' # NMHC
+#' ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
+#' l = 0.5, p = "heptane", x = 10)(30)
+#'
 #' V <- 0:130
 #' ef1 <- ef_hdv_speed(v = "Trucks",t = "RT", g = "<=7.5", e = "II", gr = 0,
 #' l = 0.5, p = "HC")
@@ -33,7 +100,8 @@
 #' plot(efs, xlab = "age")
 #' lines(efs, type = "l")
 #' }
-ef_hdv_speed <- function(v, t, g, eu, gr = 0, l = 0.5 ,p, k=1, show.equation=TRUE){
+ef_hdv_speed <- function(v, t, g, eu, x, gr = 0, l = 0.5 ,p, k=1,
+                         show.equation = FALSE){
   ef_hdv <- sysdata[[2]]
   df <- ef_hdv[ef_hdv$VEH == v &
                  ef_hdv$TYPE == t &
@@ -42,23 +110,39 @@ ef_hdv_speed <- function(v, t, g, eu, gr = 0, l = 0.5 ,p, k=1, show.equation=TRU
                  ef_hdv$GRA == gr &
                  ef_hdv$LOAD == l &
                  ef_hdv$POLLUTANT == p, ]
-  lista <- list(a = df$a,
-                b = df$b,
-                c = df$c,
-                d = df$d,
-                e = df$e,
-                Equation = paste0("(",as.character(df$Y), ")", "*", k))
   if (show.equation == TRUE) {
-    print(lista)
+    cat(paste0("a = ", df$a,
+               ", b = ", df$b,
+               ", c = ", df$c,
+               ", d = ", df$d,
+               ", e = ", df$e,
+               ", f = ", df$f, "\n"))
+    cat(paste0("Equation = ", "(",as.character(df$Y), ")", "*", k))
   }
-  f1 <- function(V){
-    a <- df$a;
-    b <- df$b;
-    c <- df$c ;
-    d <- df$d;
-    e <- df$e
-    V <- ifelse(V<df$MINV,df$MINV,ifelse(V>df$MAXV,df$MAXV,V))
-    eval(parse(text = paste0("(",as.character(df$Y), ")", "*", k)))
+  if(p %in% c("SO2","Pb")){
+    f1 <- function(V){
+      a <- df$a
+      b <- df$b
+      c <- df$c
+      d <- df$d
+      e <- df$e
+      f <- df$f
+      x <- x
+      V <- ifelse(V < df$MINV, df$MINV,
+                  ifelse(V > df$MAXV, df$MAXV, V))
+      eval(parse(text = paste0("(",as.character(df$Y), ")", "*", k)))
+    }
+  } else {
+    f1 <- function(V){
+      a <- df$a
+      b <- df$b
+      c <- df$c
+      d <- df$d
+      e <- df$e
+      f <- df$f
+      V <- ifelse(V<df$MINV,df$MINV,ifelse(V>df$MAXV,df$MAXV,V))
+      eval(parse(text = paste0("(",as.character(df$Y), ")", "*", k)))
+    }
   }
   return(f1)
 }