+++ R/emis_paved.R
@@ -9,7 +9,8 @@
 #' @param veh Numeric vector with length of elements equals to number of streets
 #' It is an array with dimenssions number of streets x hours of day x days of week
 #' @param lkm Length of each link
-#' @param k K_PM30 = 3.23, K_PM15 = 0.77, K_PM10 = 0.62 and K_PM2.5 = 0.15
+#' @param k K_PM30 = 3.23 (g/vkm), K_PM15 = 0.77 (g/vkm), K_PM10 = 0.62 (g/vkm)
+#' and K_PM2.5 = 0.15  (g/vkm).
 #' @param sL1 Silt loading (g/m2) for roads with ADT <= 500
 #' @param sL2 Silt loading (g/m2) for roads with ADT > 500 and <= 5000
 #' @param sL3 Silt loading (g/m2) for roads with ADT > 5000 and <= 1000
@@ -33,20 +34,26 @@
 #' class(emi)
 #' head(emi)
 #' }
-emis_paved <- function(veh, lkm, k, sL1, sL2, sL3, sL4, W) {
-  message("Estimation of dry hours only, aggregation should include rainy hours")
-  if (class(veh)!="array" | class(W)!="array") {
-    stop("class of veh or W should be array")
-  }
-  d <- sapply(1:dim(veh)[3], function(i) rowSums(veh[,,i]))
-  sL <- sapply(1:ncol(d), function(i)
-    ifelse(d[,i]<=500, sL1,
-           ifelse(d[,i]>500 & d[,i]<=5000, sL2,
-                  ifelse(d[,i]>5000 & d[,i]<= 10000, sL3,sL4)))
-  )
-  emi <-simplify2array(lapply(1:7, function(i)
-    veh[,,i]*lkm*k*sL[,i]^0.91*W[,,i]^1.02
-  ))
+emis_paved <- function(veh,
+                       lkm,
+                       k = 0.62,   # K_PM10 = 0.62 (g/vkm)
+                       sL1 = 0.6,  # g/m^2
+                       sL2 = 0.2,  # g/m^2
+                       sL3 = 0.06, # g/m^2
+                       sL4 = 0.03, # g/m^2
+                       W) {
+  sL <- array(data = ifelse(veh <= 500, sL1,
+                            ifelse(veh > 500 & veh <= 5000, sL2,
+                                   ifelse(veh > 5000 & veh <=1000, sL3,
+                                          sL4))),
+              dim = dim(veh))
+  lkm <- array(lkm, dim = dim(veh))
+  k <- array(k, dim = dim(veh))
+  emi <- veh * lkm * k * sL^0.91 * W^1.02
   emi[is.na(emi)] <- 0
-  return(EmissionsArray(emi))
+  if(length(dim(emi)) == 2){
+    return(Emissions(emi))
+  } else {
+    return(EmissionsArray(emi))
+  }
 }